Extracting methods into common class
