+++ Exebite.GoogleSheetAPI/Connectors/Restaurants/Base/RestaurantConnector.cs
@@ -1,7 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Threading;
+using Either;
+using Exebite.DataAccess.Repositories;
 using Exebite.DomainModel;
+using Exebite.GoogleSheetAPI.Common;
 using Exebite.GoogleSheetAPI.SheetExtractor;
 using Google.Apis.Sheets.v4.Data;
 
@@ -10,10 +14,16 @@ namespace Exebite.GoogleSheetAPI.Connectors.Restaurants.Base
     public abstract class RestaurantConnector : IRestaurantConnector
     {
         private readonly string _kreditTab = "Kredit";
+        private readonly IRestaurantQueryRepository _restaurantQueryRepository;
 
-        protected RestaurantConnector(IGoogleSheetExtractor googleSheetService)
+        protected RestaurantConnector(
+            IGoogleSheetExtractor googleSheetService,
+            IRestaurantQueryRepository restaurantQueryRepository,
+            string restaurantName)
         {
+            _restaurantQueryRepository = restaurantQueryRepository;
             GoogleSheetService = googleSheetService;
+            Restaurant = GetRestaurant(restaurantName);
         }
 
         internal IGoogleSheetExtractor GoogleSheetService { get; set; }
@@ -404,6 +414,93 @@ namespace Exebite.GoogleSheetAPI.Connectors.Restaurants.Base
             }
 
             return monthLocal;
+        }
+
+        /// <summary>
+        /// Finds the merge inside all the sheets.
+        /// </summary>
+        /// <param name="providedDate">Date for which needs to be found date range</param>
+        /// <returns>Result with all <seealso cref="MergedRegion"/></returns>
+        internal Result<MergedRegion> FindDateRangeInSheets(
+            DateTime providedDate)
+        {
+            var foundRegion = GoogleSheetService.GetWorkSheets(SheetId)
+                .Select(SheetMerges)
+                .Select(mergeList => GetMergeWithDate(mergeList, providedDate))
+                .Where(result => result.IsSuccess)
+                .FirstOrDefault();
+
+            if (foundRegion != null)
+            {
+                return foundRegion;
+            }
+
+            return Result<MergedRegion>.Fail(null, string.Format("No merge with provided date {0}.", providedDate.ToString()));
+        }
+
+        /// <summary>
+        /// From all the merges, find the one that has provided date.
+        /// </summary>
+        /// <param name="merges">Collection of all <seealso cref="MergedRegion"/></param>
+        /// <param name="providedDate">Date for which <paramref name="merges"/> will be checked.</param>
+        /// <returns>Result with <seealso cref="MergedRegion"/></returns>
+        internal Result<MergedRegion> GetMergeWithDate(
+            IEnumerable<MergedRegion> merges,
+            DateTime providedDate)
+        {
+            foreach (var merge in merges)
+            {
+                var result = GoogleSheetService.ReadDateTime(merge.A1FirstCell, SheetId);
+                if (result.IsSuccess)
+                {
+                    DateTime parsedDate = result.Value;
+
+                    if (providedDate.Year != parsedDate.Year
+                        || providedDate.Month != parsedDate.Month)
+                    {
+                        break;
+                    }
+
+                    if (providedDate.Date.Equals(parsedDate.Date))
+                    {
+                        return Result<MergedRegion>.Success(merge);
+                    }
+                }
+
+                // There is a limitation on Google Side for 100 calls per second.
+                // We have added this thread sleep to avoid such issues.
+                Thread.Sleep(Constants.SLEEP_TIME);
+            }
+
+            return Result<MergedRegion>.Fail(null, "Sheet doesn't contain this date.");
+        }
+
+        /// <summary>
+        /// Returns a list of all merged regions
+        /// </summary>
+        /// <param name="sheet">Google sheet for which will be checked for merged regions.</param>
+        /// <returns>Returns all merged regions from the sheet.</returns>
+        private IEnumerable<MergedRegion> SheetMerges(Sheet sheet)
+        {
+            if (sheet.Merges == null)
+            {
+                return new List<MergedRegion>();
+            }
+
+            return sheet.Merges.Select(merge => new MergedRegion(sheet, merge));
+        }
+
+        /// <summary>
+        /// Get the restaurant from the database
+        /// </summary>
+        /// <param name="name">Name of the restaurant that needs to be returned from database.</param>
+        /// <returns>Returns restaurant object from database if exists, otherwise null.</returns>
+        private Restaurant GetRestaurant(string name)
+        {
+            return _restaurantQueryRepository
+                    .Query(new RestaurantQueryModel { Name = name })
+                    .Map(res => res.Items.FirstOrDefault())
+                    .Reduce(r => null, ex => Console.WriteLine(ex.ToString()));
         }
     }
 }
