+++ Exebite.GoogleSheetAPI/Connectors/Restaurants/LipaConnector.cs
@@ -1,36 +1,25 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
-using Either;
 using Exebite.DataAccess.Repositories;
 using Exebite.DomainModel;
 using Exebite.GoogleSheetAPI.Common;
 using Exebite.GoogleSheetAPI.Connectors.Restaurants.Base;
 using Exebite.GoogleSheetAPI.GoogleSSFactory;
 using Exebite.GoogleSheetAPI.SheetExtractor;
-using Google.Apis.Sheets.v4.Data;
 
 namespace Exebite.GoogleSheetAPI.Connectors.Restaurants
 {
     public sealed class LipaConnector : RestaurantConnector, ILipaConnector
     {
-        private readonly string _sheetId;
-        private readonly Restaurant _restaurant;
-        private readonly IRestaurantQueryRepository _restaurantQueryRepository;
-
         public LipaConnector(
             IGoogleSheetExtractor googleSheetService,
             IGoogleSpreadsheetIdFactory googleSSIdFactory,
             IRestaurantQueryRepository restaurantQueryRepository)
-            : base(googleSheetService)
+            : base(googleSheetService, restaurantQueryRepository, "Restoran pod Lipom")
         {
-            _restaurantQueryRepository = restaurantQueryRepository;
-            _sheetId = googleSSIdFactory.GetSheetId(Enums.ESheetOwner.LIPA);
-            _restaurant = GetRestaurant();
-            SheetId = _sheetId;
+            SheetId = googleSSIdFactory.GetSheetId(Enums.ESheetOwner.LIPA);
             DailyMenuSheet = GetLocalMonthName(DateTime.Now.Month) + DateTime.Now.Year;
-            Restaurant = _restaurant;
         }
 
         /// <summary>
@@ -71,7 +60,7 @@ namespace Exebite.GoogleSheetAPI.Connectors.Restaurants
                         endColumn: foundMerge.Value.Range.EndColumnIndex.Value,
                         endRow: 4); // End corner
 
-                var offersList = GoogleSheetService.ReadSheetData(string.Format("'{0}'!{1}", foundMerge.Value.SheetName, namesRange), _sheetId);
+                var offersList = GoogleSheetService.ReadSheetData(string.Format("'{0}'!{1}", foundMerge.Value.SheetName, namesRange), SheetId);
 
                 var result = new List<Food>();
                 var foodNames = offersList.Values.First().ToList();
@@ -86,7 +75,7 @@ namespace Exebite.GoogleSheetAPI.Connectors.Restaurants
                         !string.IsNullOrWhiteSpace(foodPrice) &&
                         decimal.TryParse(foodPrice, out decimal price))
                     {
-                        result.Add(new Food { Name = foodName, Price = price, RestaurantId = _restaurant.Id });
+                        result.Add(new Food { Name = foodName, Price = price, RestaurantId = Restaurant.Id });
                         if (i == foodNames.Count - 1)
                         {
                             // currently in Lipa restaurant SOUP is always the last meal in daily sheet
@@ -101,89 +90,6 @@ namespace Exebite.GoogleSheetAPI.Connectors.Restaurants
             {
                 return new List<Food>();
             }
-        }
-
-        /// <summary>
-        /// Get the restaurant from the database
-        /// </summary>
-        /// <returns>Returns restaurant object from database if exists, otherwise null.</returns>
-        private Restaurant GetRestaurant()
-        {
-            return _restaurantQueryRepository
-                    .Query(new RestaurantQueryModel { Name = "Restoran pod Lipom" })
-                    .Map(res => res.Items.FirstOrDefault())
-                    .Reduce(r => null, ex => Console.WriteLine(ex.ToString()));
-        }
-
-        /// <summary>
-        /// Returns a list of all merged regions
-        /// </summary>
-        /// <param name="sheet">Google sheet for which will be checked for merged regions.</param>
-        /// <returns>Returns all merged regions from the sheet.</returns>
-        private IEnumerable<MergedRegion> SheetMerges(Sheet sheet)
-        {
-            if (sheet.Merges == null)
-            {
-                return new List<MergedRegion>();
-            }
-
-            return sheet.Merges.Select(merge => new MergedRegion(sheet, merge));
-        }
-
-        /// <summary>
-        /// Finds the merge inside all the sheets.
-        /// </summary>
-        /// <param name="providedDate">Date for which needs to be found date range</param>
-        /// <returns>Result with all <seealso cref="MergedRegion"/></returns>
-        private Result<MergedRegion> FindDateRangeInSheets(DateTime providedDate)
-        {
-            var foundRegion = GoogleSheetService.GetWorkSheets(_sheetId)
-                .Select(SheetMerges)
-                .Select(mergeList => GetMergeWithDate(mergeList, providedDate))
-                .Where(result => result.IsSuccess)
-                .FirstOrDefault();
-
-            if (foundRegion != null)
-            {
-                return foundRegion;
-            }
-
-            return Result<MergedRegion>.Fail(null, string.Format("No merge with provided date {0}.", providedDate.ToString()));
-        }
-
-        /// <summary>
-        /// From all the merges, find the one that has provided date.
-        /// </summary>
-        /// <param name="merges">Collection of all <seealso cref="MergedRegion"/></param>
-        /// <param name="providedDate">Date for which <paramref name="merges"/> will be checked.</param>
-        /// <returns>Result with <seealso cref="MergedRegion"/></returns>
-        private Result<MergedRegion> GetMergeWithDate(IEnumerable<MergedRegion> merges, DateTime providedDate)
-        {
-            foreach (var merge in merges)
-            {
-                var result = GoogleSheetService.ReadDateTime(merge.A1FirstCell, _sheetId);
-                if (result.IsSuccess)
-                {
-                    DateTime parsedDate = result.Value;
-
-                    if (providedDate.Year != parsedDate.Year
-                        || providedDate.Month != parsedDate.Month)
-                    {
-                        break;
-                    }
-
-                    if (providedDate.Date.Equals(parsedDate.Date))
-                    {
-                        return Result<MergedRegion>.Success(merge);
-                    }
-                }
-
-                // There is a limitation on Google Side for 100 calls per second.
-                // We have added this thread sleep to avoid such issues.
-                Thread.Sleep(Constants.SLEEP_TIME);
-            }
-
-            return Result<MergedRegion>.Fail(null, "Sheet doesn't contain this date.");
         }
     }
 }
